ticket:193 : jamesmanners : monsoon3 build config including ksh to bash translation : task : scripts : minor
git-svn-id: https://metomi/svn/socrates.xm/main/trunk@1936 7a691fcb-5d1e-4fc1-8b6b-54125fa6be09

+++ examples/cosp/check_cosp_github.sh
@@ -7,8 +7,13 @@
 rm -rf COSPv2.0-master
 rm -f master.zip
 wget -nc https://github.com/CFMIP/COSPv2.0/archive/refs/heads/master.zip
-unzip master.zip
-cp COSPv2.0-master/driver/data/inputs/UKMO/cosp_input_um_2d.nc .
+if [ -f master.zip ] ; then
+  unzip master.zip
+  cp COSPv2.0-master/driver/data/inputs/UKMO/cosp_input_um_2d.nc .
+else
+  echo 'COSP github not accessible'
+  exit 0
+fi
 
 # Now check the diff to the current fcm version
 ierr=0
